+++ r4/src/test/java/io/github/jy95/fds/r4/translators/DurationDurationMaxTest.java
@@ -1,33 +1,20 @@
 package io.github.jy95.fds.r4.translators;
 
-import io.github.jy95.fds.r4.DosageAPIR4;
-import io.github.jy95.fds.r4.AbstractFhirTest;
 import io.github.jy95.fds.common.types.DisplayOrder;
+import io.github.jy95.fds.common.types.DosageAPI;
+import io.github.jy95.fds.r4.DosageAPIR4;
+import io.github.jy95.fds.r4.config.FDSConfigR4;
+import io.github.jy95.fds.translators.AbstractDurationDurationMaxTest;
 import org.hl7.fhir.r4.model.Dosage;
 import org.hl7.fhir.r4.model.Timing;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.MethodSource;
 
+import java.util.List;
 import java.util.Locale;
-import java.util.concurrent.ExecutionException;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
+public class DurationDurationMaxTest extends AbstractDurationDurationMaxTest<FDSConfigR4, Dosage> {
 
-public class DurationDurationMaxTest extends AbstractFhirTest {
-
-    @ParameterizedTest
-    @MethodSource("localeProvider")
-    void testNoDuration(Locale locale) throws ExecutionException, InterruptedException {
-        Dosage dosage = new Dosage();
-        DosageAPIR4 dosageUtils = DayOfWeekTest.getDosageAPI(locale, DisplayOrder.DURATION_DURATION_MAX);
-        String result = dosageUtils.asHumanReadableText(dosage).get();
-        assertEquals("", result);
-    }
-
-    @ParameterizedTest
-    @MethodSource("localeProvider")
-    void testWithDurationOnly(Locale locale) throws ExecutionException, InterruptedException {
-        DosageAPIR4 dosageUtils = DayOfWeekTest.getDosageAPI(locale, DisplayOrder.DURATION_DURATION_MAX);
+    @Override
+    protected Dosage generateWithDurationOnly() {
         Dosage dosage = new Dosage();
         Timing timing = new Timing();
         Timing.TimingRepeatComponent repeatComponent = new Timing.TimingRepeatComponent();
@@ -35,16 +22,11 @@ public class DurationDurationMaxTest extends AbstractFhirTest {
         repeatComponent.setDurationUnit(Timing.UnitsOfTime.D);
         timing.setRepeat(repeatComponent);
         dosage.setTiming(timing);
-
-        String result = dosageUtils.asHumanReadableText(dosage).get();
-        String expectedResult = getExpectedText1(locale);
-        assertEquals(expectedResult, result);
+        return dosage;
     }
 
-    @ParameterizedTest
-    @MethodSource("localeProvider")
-    void testWithDurationMaxOnly(Locale locale) throws ExecutionException, InterruptedException {
-        DosageAPIR4 dosageUtils = DayOfWeekTest.getDosageAPI(locale, DisplayOrder.DURATION_DURATION_MAX);
+    @Override
+    protected Dosage generateWithDurationMaxOnly() {
         Dosage dosage = new Dosage();
         Timing timing = new Timing();
         Timing.TimingRepeatComponent repeatComponent = new Timing.TimingRepeatComponent();
@@ -52,16 +34,11 @@ public class DurationDurationMaxTest extends AbstractFhirTest {
         repeatComponent.setDurationUnit(Timing.UnitsOfTime.D);
         timing.setRepeat(repeatComponent);
         dosage.setTiming(timing);
-
-        String result = dosageUtils.asHumanReadableText(dosage).get();
-        String expectedResult = getExpectedText2(locale);
-        assertEquals(expectedResult, result);
+        return dosage;
     }
 
-    @ParameterizedTest
-    @MethodSource("localeProvider")
-    void testWithBothDuration(Locale locale) throws ExecutionException, InterruptedException {
-        DosageAPIR4 dosageUtils = DayOfWeekTest.getDosageAPI(locale, DisplayOrder.DURATION_DURATION_MAX);
+    @Override
+    protected Dosage generateWithBothDuration() {
         Dosage dosage = new Dosage();
         Timing timing = new Timing();
         Timing.TimingRepeatComponent repeatComponent = new Timing.TimingRepeatComponent();
@@ -70,45 +47,24 @@ public class DurationDurationMaxTest extends AbstractFhirTest {
         repeatComponent.setDurationUnit(Timing.UnitsOfTime.D);
         timing.setRepeat(repeatComponent);
         dosage.setTiming(timing);
-
-        String result = dosageUtils.asHumanReadableText(dosage).get();
-        String expectedResult = getExpectedText3(locale);
-        assertEquals(expectedResult, result);
+        return dosage;
     }
 
-    private String getExpectedText1(Locale locale) {
-        if (locale.equals(Locale.ENGLISH)) {
-            return "over 3 days";
-        } else if (locale.equals(Locale.FRENCH)) {
-            return "durant 3 jours";
-        } else if (locale.equals(Locale.GERMAN)) {
-            return "für 3 Tage";
-        } else {
-            return "gedurende 3 dagen";
-        }
+    @Override
+    public DosageAPI<FDSConfigR4, Dosage> getDosageAPI(Locale locale, DisplayOrder displayOrder) {
+        return new DosageAPIR4(FDSConfigR4.builder()
+                .displayOrder(List.of(displayOrder))
+                .locale(locale)
+                .build());
     }
 
-    private String getExpectedText2(Locale locale) {
-        if (locale.equals(Locale.ENGLISH)) {
-            return "maximum 3 days";
-        } else if (locale.equals(Locale.FRENCH)) {
-            return "maximum 3 jours";
-        } else if (locale.equals(Locale.GERMAN)) {
-            return "maximal 3 Tage";
-        } else {
-            return "maximaal 3 dagen";
-        }
+    @Override
+    public DosageAPI<FDSConfigR4, Dosage> getDosageAPI(FDSConfigR4 config) {
+        return new DosageAPIR4(config);
     }
 
-    private String getExpectedText3(Locale locale) {
-        if (locale.equals(Locale.ENGLISH)) {
-            return "over 3 days ( maximum 5 days )";
-        } else if (locale.equals(Locale.FRENCH)) {
-            return "durant 3 jours ( maximum 5 jours )";
-        } else if (locale.equals(Locale.GERMAN)) {
-            return "für 3 Tage ( maximal 5 Tage )";
-        } else {
-            return "gedurende 3 dagen ( maximaal 5 dagen )";
-        }
+    @Override
+    public Dosage generateEmptyDosage() {
+        return new Dosage();
     }
 }
